Staking UI fixes
